requirements.txt
asgiref==3.9.1
Django==5.2.5
django-htmx==1.23.2
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.3
django-extensions==4.1